jtl-karrio/source.requirements.insiders.txt
--extra-index-url https://karrio.gateway.scarf.sh/simple/
asgiref==3.9.1
attrs==25.3.0
backoff==2.2.1
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
croniter==6.0.0
cryptography==46.0.1
cssselect2==0.8.0
deprecation==2.1.0
diff-match-patch==20241021
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.6
django-auditlog==3.2.1
django-constance==4.3.2
django-cors-headers==4.9.0
django-downloadview==2.4.0
django-email-verification==0.3.3
django-extensions==4.1
django-filter==25.1
django-formtools==2.5.1
django-health-check==3.20.0
django-import-export==4.3.9
django-oauth-toolkit==3.0.1
django-organizations==2.5.0
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-picklefield==3.3
django-redis==6.0.0
django-two-factor-auth==1.18.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-api-tracking==1.8.4
drf-spectacular==0.28.0
fonttools==4.60.0
googleapis-common-protos==1.70.0
graphql-core==3.2.6
grpcio==1.75.0
gunicorn==23.0.0
h11==0.16.0
hiredis==3.2.1
huey==2.5.3
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
jsonfield==3.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jstruct==2021.11
jwcrypto==1.5.6
karrio @ file://${PWD}/modules/sdk
karrio_australiapost @ file://${PWD}/modules/connectors/australiapost
karrio_canadapost @ file://${PWD}/modules/connectors/canadapost
karrio_chronopost @ file://${PWD}/community/plugins/chronopost
karrio_dhl_express @ file://${PWD}/modules/connectors/dhl_express
karrio_dhl_parcel_de @ file://${PWD}/modules/connectors/dhl_parcel_de
karrio_dhl_poland @ file://${PWD}/modules/connectors/dhl_poland
karrio_dhl_universal @ file://${PWD}/modules/connectors/dhl_universal
karrio_dpd @ file://${PWD}/community/plugins/dpd
karrio_dtdc @ file://${PWD}/ee/insiders/modules/connectors/dtdc
karrio_easypost @ file://${PWD}/community/plugins/easypost
karrio_easyship @ file://${PWD}/community/plugins/easyship
karrio_eshipper @ file://${PWD}/community/plugins/eshipper
karrio_fedex @ file://${PWD}/modules/connectors/fedex
karrio_freightcom @ file://${PWD}/community/plugins/freightcom
karrio_generic @ file://${PWD}/modules/connectors/generic
karrio_laposte @ file://${PWD}/modules/connectors/laposte
karrio_purolator @ file://${PWD}/modules/connectors/purolator
karrio_sapient @ file://${PWD}/community/plugins/sapient
karrio_seko @ file://${PWD}/modules/connectors/seko
karrio_sendle @ file://${PWD}/modules/connectors/sendle
karrio_server @ file://${PWD}/apps/api
karrio_server_admin @ file://${PWD}/ee/insiders/modules/admin
karrio_server_apps @ file://${PWD}/ee/insiders/modules/apps
karrio_server_audit @ file://${PWD}/ee/insiders/modules/audit
karrio_server_automation @ file://${PWD}/ee/insiders/modules/automation
karrio_server_cloud @ file://${PWD}/ee/insiders/modules/cloud
karrio_server_core @ file://${PWD}/modules/core
karrio_server_data @ file://${PWD}/modules/data
karrio_server_documents @ file://${PWD}/modules/documents
karrio_server_events @ file://${PWD}/modules/events
karrio_server_graph @ file://${PWD}/modules/graph
karrio_server_manager @ file://${PWD}/modules/manager
karrio_server_orders @ file://${PWD}/modules/orders
karrio_server_orgs @ file://${PWD}/ee/insiders/modules/orgs
karrio_server_pricing @ file://${PWD}/modules/pricing
karrio_server_proxy @ file://${PWD}/modules/proxy
karrio_ups @ file://${PWD}/modules/connectors/ups
karrio_usps @ file://${PWD}/modules/connectors/usps
karrio_usps_international @ file://${PWD}/modules/connectors/usps_international
lia-web==0.2.3
lxml==4.9.4
lxml-stubs==0.5.1
MarkupSafe==3.0.2
more-itertools==10.8.0
oauthlib==3.3.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-dbapi==0.58b0
opentelemetry-instrumentation-django==0.58b0
opentelemetry-instrumentation-logging==0.58b0
opentelemetry-instrumentation-psycopg2==0.58b0
opentelemetry-instrumentation-redis==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-wsgi==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
packaging==25.0
phonenumbers==9.0.14
pillow==11.3.0
posthog==6.7.6
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.10
py-soap==2020.3.2
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
PyPDF2==3.0.1
pyphen==0.17.2
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
pyzint==0.1.10
qrcode==8.2
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sentry-sdk==2.38.0
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.282.0
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.36.0
validators==0.35.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
wrapt==1.17.3
xmltodict==1.0.2
zipp==3.23.0
zopfli==0.2.3.post1

